refactor: divide codes into three Go files (webdriver, browser, element)
